server and client code refactoring
