+++ apps/client/index.html
@@ -2,7 +2,7 @@
 <html lang="en">
   <head>
     <meta charset="UTF-8" />
-    <link rel="icon" type="image/svg+xml" href="./public/AuroraGrille.svg" />
+    <link rel="icon" type="image/svg+xml" href="./AuroraGrille.svg" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Aurora Grille || POS System</title>
   </head>
